Backend/data_agent.py
import os
import requests
from bs4 import BeautifulSoup
import google.generativeai as genai
import logging

# Enable logging
logging.basicConfig(level=logging.INFO)

class DataAgent:
    def __init__(self, api_key):
        self.api_key = api_key
        genai.configure(api_key=self.api_key)
        self.model = genai.GenerativeModel('gemini-pro')

    def scrape_data(self, url):
        """
        Scrapes data from a given URL.
        """
        try:
            response = requests.get(url)
            response.raise_for_status()  # Raise an exception for bad status codes
            soup = BeautifulSoup(response.text, 'html.parser')
            return soup.get_text()
        except requests.exceptions.RequestException as e:
            logging.error(f"Error scraping data from {url}: {e}")
            return None

    def get_data_for_district(self, district, date):
        """
        Fetches and processes data for a specific district and date.
        This is a placeholder and should be replaced with actual scraping and processing logic.
        """
        logging.info(f"Fetching data for {district} on {date}...")

        # Example: Scrape a relevant website (replace with a real URL)
        scraped_text = self.scrape_data(f"https://www.example.com/weather/{district}/{date}")

        if not scraped_text:
            return None

        # Use Gemini to extract structured data
        prompt = f"""
        From the following text, extract the following information for {district} on {date}:
        - Temperature (°C)
        - Humidity (%)
        - Wind Speed (km/h)

        Text:
        {scraped_text}
        """

        try:
            response = self.model.generate_content(prompt)
            # Process the response to extract the data
            # This will depend on the format of the Gemini API response
            processed_data = self.process_gemini_response(response)
            return processed_data
        except Exception as e:
            logging.error(f"Error processing data with Gemini: {e}")
            return None

    def process_gemini_response(self, response):
        """
        Processes the response from the Gemini API to extract structured data.
        This is a placeholder and needs to be implemented based on the actual response format.
        """
        # Example implementation
        # The actual implementation will depend on the structure of the response from Gemini
        # For now, we'll return a dummy dictionary
        return {
            "temperature": 25,
            "humidity": 60,
            "wind_speed": 15,
        }

if __name__ == '__main__':
    # Example usage
    gemini_api_key = os.environ.get("AIzaSyAvwkP3u0wvx5sDOYNUWaPRAMbe6HDtWmw")
    if not gemini_api_key:
        print("Please set the GEMINI_API_KEY environment variable.")
    else:
        agent = DataAgent(api_key=gemini_api_key)
        data = agent.get_data_for_district("Kolkata", "2025-04-15")
        if data:
            print("Successfully fetched and processed data:")
            print(data)