Backend/requirements.txt
Flask
pandas
prophet
flask-cors
google-generativeai
beautifulsoup4
requests